Add Tag Model/Controller/View/Function

+++ app/controllers/books_controller.rb
@@ -12,7 +12,7 @@ class BooksController < ApplicationController
 		@book = Book.find(params[:id])
 
 		# カテゴリPOST処理
-		if params.require(:ctgname)
+		if params[:sendctg]
 			@category = Category.find_by(name: params.require(:ctgname))
 			# 既存のカテゴリの場合
 			if @category
@@ -31,14 +31,44 @@ class BooksController < ApplicationController
 				end
 			end
 		end
+		# タグPOST処理
+		if params[:sendtag]
+			@tag = Tag.find_by(name: params.require(:tagname))
+			# 既存のタグの場合
+			if @tag
+				@tagrelation = Tagrelation.new(book_id: @book.id, tag_id: @tag.id)
+				if @tagrelation.save
+					redirect_to book_path(@book.id)
+				end
+			# 新規タグの場合
+			else
+				@tag = Tag.new(name: params.require(:tagname))
+				if @tag.save
+					@tagrelation = Tagrelation.new(book_id: @book.id, tag_id: @tag.id)
+					if @tagrelation.save
+						redirect_to book_path(@book.id)
+					end
+				end
+			end
+		end
 	end
 
 	def delatr
-		bookid = params[:id]
-		@category = Category.find_by(name: params.require(:ctgname))
-		@ctgrelation = Ctgrelation.find_by(book_id: bookid, category_id: @category.id)
-		if @ctgrelation.destroy
-			redirect_to book_path(bookid)
+		# カテゴリDELETE処理
+		if params[:ctgname]
+			bookid = params[:id]
+			@category = Category.find_by(name: params.require(:ctgname))
+			@ctgrelation = Ctgrelation.find_by(book_id: bookid, category_id: @category.id)
+			if @ctgrelation.destroy
+				redirect_to book_path(bookid)
+			end
+		elsif params[:tagname]
+			bookid = params[:id]
+			@tag = Tag.find_by(name: params.require(:tagname))
+			@tagrelation = Tagrelation.find_by(book_id: bookid, tag_id: @tag.id)
+			if @tagrelation.destroy
+				redirect_to book_path(bookid)
+			end
 		end
 	end
 end
